requirements.txt
transformers
transformers[sentencepiece]
datasets
sacrebleu
rouge_score
py7zr
pandas
nltk
tqdm
PyYAML
matplotlib
torch
notebook
boto3
mypy-boto3-s3
ensure
fastapi
uvicorn
Jinja2
python-box[all]~=7.0
-e .